feat(memory-stack-state-owned): modify to own the backend type
requires using our patch of the upstream evm until https://github.com/rust-blockchain/evm/pull/62 is merged
