+++ Cargo.toml
@@ -14,3 +14,7 @@ opt-level = "z"
 lto = true
 codegen-units = 1
 panic = "abort"
+
+
+[patch."https://github.com/rust-blockchain/evm"]
+evm = { git = "https://github.com/gakonst/evm", branch = "chore/expose-stack-module" }
